// src/clients/CatalogAdministration/models/CatalogAdminsSearchMerchandizingRule.ts
export interface CatalogAdminsSearchMerchandizingRule {
    /**
     * Unique identifier. Will be generated if not provided.
     * @type {string}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    description?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    tags?: Array<string> | null;
    /**
     * 
     * @type {CatalogAdminsSearchMerchandizingRuleContext}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    context?: CatalogAdminsSearchMerchandizingRuleContext;
    /**
     * 
     * @type {SearchMerchandizingRuleImpact}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    ruleImpact?: SearchMerchandizingRuleImpact;
    /**
     * 
     * @type {AdminUserAuditInfo}
     * @memberof CatalogAdminsSearchMerchandizingRule
     */
    auditInfo?: AdminUserAuditInfo;
}