// src/clients/CatalogAdministration/models/SearchMerchandizingRuleCollection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kibo Catalog Admin Services
 * OpenAPI Spec for Kibo Catalog Admin Services
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { SearchMerchandizingRule } from './SearchMerchandizingRule';
import {
    SearchMerchandizingRuleFromJSON,
    SearchMerchandizingRuleFromJSONTyped,
    SearchMerchandizingRuleToJSON,
} from './SearchMerchandizingRule';

/**
 * 
 * @export
 * @interface SearchMerchandizingRuleCollection
 */
export interface SearchMerchandizingRuleCollection {
    /**
     * 
     * @type {number}
     * @memberof SearchMerchandizingRuleCollection
     */
    startIndex?: number;
    /**
     * 
     * @type {number}
     * @memberof SearchMerchandizingRuleCollection
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof SearchMerchandizingRuleCollection
     */
    pageCount?: number;
    /**
     * 
     * @type {number}
     * @memberof SearchMerchandizingRuleCollection
     */
    totalCount?: number;
    /**
     * 
     * @type {Array<SearchMerchandizingRule>}
     * @memberof SearchMerchandizingRuleCollection
     */
    items?: Array<SearchMerchandizingRule> | null;
}

/**
 * Check if a given object implements the SearchMerchandizingRuleCollection interface.
 */
export function instanceOfSearchMerchandizingRuleCollection(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SearchMerchandizingRuleCollectionFromJSON(json: any): SearchMerchandizingRuleCollection {
    return SearchMerchandizingRuleCollectionFromJSONTyped(json, false);
}

export function SearchMerchandizingRuleCollectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchMerchandizingRuleCollection {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'startIndex': !exists(json, 'startIndex') ? undefined : json['startIndex'],
        'pageSize': !exists(json, 'pageSize') ? undefined : json['pageSize'],
        'pageCount': !exists(json, 'pageCount') ? undefined : json['pageCount'],
        'totalCount': !exists(json, 'totalCount') ? undefined : json['totalCount'],
        'items': !exists(json, 'items') ? undefined : (json['items'] === null ? null : (json['items'] as Array<any>).map(SearchMerchandizingRuleFromJSON)),
    };
}

export function SearchMerchandizingRuleCollectionToJSON(value?: SearchMerchandizingRuleCollection | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'startIndex': value.startIndex,
        'pageSize': value.pageSize,
        'pageCount': value.pageCount,
        'totalCount': value.totalCount,
        'items': value.items === undefined ? undefined : (value.items === null ? null : (value.items as Array<any>).map(SearchMerchandizingRuleToJSON)),
    };
}